Add claims

+++ src/resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt
 
 from src.models.item import ItemModel
 
@@ -86,16 +86,22 @@ class Item(Resource):
     @jwt_required()
     def delete(name: str) -> tuple:
         """
-        Deletes given object.
+        Deletes given object if existing.
 
         :param name: String name.
-        :return: {'message': 'Item deleted!'}
+        :return: {'message': String}
         """
+        claims = get_jwt()
+
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required!'}, 401
+
         item = ItemModel.find_by_name(name)
 
         if item:
             item.delete_from_db()
-        return {'message': 'Item deleted!'}, 200
+            return {'message': 'Item deleted!'}, 200
+        return {'message': 'Item not found!'}, 404
 
 
 class ItemList(Resource):
